+++ ts/apps/polaris-cli/src/app/tasks/index.ts
@@ -1,3 +1,4 @@
+export * from './promise-task';
 export * from './run-process-task';
 export * from './task-executor';
 export * from './task';