+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Local npm registry using [Verdaccio](https://verdaccio.org/)
 * Run e2e tests in GithubActions
 * Docker-build command using an Nx executor
+* Generate a Visual Studio Code .devcontainer configuration when creating a new Polaris workspace